Move sources to content dir
