--- assets/externalized/script-records-RPCs.json
@@ -1,6 +1,0 @@
-[
-    {
-        "profile": "",
-        "records": []
-    }
-]
