+++ src/externalized/content/CMakeLists.txt
@@ -4,6 +4,7 @@ set(JA2_SOURCES
     ${LOCAL_JA2_HEADERS}
     ${CMAKE_CURRENT_SOURCE_DIR}/ContentMercs.cc
     ${CMAKE_CURRENT_SOURCE_DIR}/Dialogs.cc
+    ${CMAKE_CURRENT_SOURCE_DIR}/NPCQuoteInfo.cc
     PARENT_SCOPE
 )
 set(JA2_INCLUDES
