+++ src/externalized/content/NPCQuoteInfo.h
