--- src/externalized/tacticalai/CMakeLists.txt
@@ -1,12 +1,0 @@
-file(GLOB LOCAL_JA2_HEADERS ${CMAKE_CURRENT_SOURCE_DIR}/*.h)
-set(JA2_SOURCES
-    ${JA2_SOURCES}
-    ${LOCAL_JA2_HEADERS}
-    ${CMAKE_CURRENT_SOURCE_DIR}/NPCQuoteInfo.cc
-    PARENT_SCOPE
-)
-set(JA2_INCLUDES
-    ${JA2_INCLUDES}
-    ${CMAKE_CURRENT_SOURCE_DIR}
-    PARENT_SCOPE
-)
